[refactor] Move primitive_types.py to types folder.
[ghstack-poisoned]
