+++ python/taichi/core/__init__.py
@@ -1,6 +1,9 @@
-from taichi.core.primitive_types import *
 from taichi.core.settings import get_os_name
 from taichi.core.util import *
+# TODO: move this to taichi/__init__.py.
+#       This is blocked since we currently require import this before taichi.lang
+#       but yapf refuse to give up formatting there.
+from taichi.types import *
 
 ti_core.build = build
 ti_core.load_module = load_module
