+++ python/taichi/lang/kernel_arguments.py
@@ -1,4 +1,3 @@
-from taichi.core.primitive_types import u64
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang.any_array import AnyArray
 from taichi.lang.enums import Layout
@@ -6,6 +5,7 @@ from taichi.lang.expr import Expr
 from taichi.lang.snode import SNode
 from taichi.lang.sparse_matrix import SparseMatrixBuilder
 from taichi.lang.util import cook_dtype, to_taichi_type
+from taichi.types.primitive_types import u64
 
 
 def ext_arr():
