+++ python/taichi/lang/kernel_impl.py
@@ -5,7 +5,6 @@ import inspect
 import re
 
 import numpy as np
-from taichi.core import primitive_types
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang import impl, util
 from taichi.lang.ast_checker import KernelSimplicityASTChecker
@@ -16,6 +15,7 @@ from taichi.lang.ndarray import Ndarray
 from taichi.lang.shell import _shell_pop_print, oinspect
 from taichi.lang.transformer import ASTTransformerTotal
 from taichi.misc.util import obsolete
+from taichi.types import primitive_types
 
 import taichi as ti
