+++ python/taichi/types/__init__.py
@@ -1,0 +1,1 @@
+from .primitive_types import *
