+++ python/taichi/ui/staging_buffer.py
@@ -1,11 +1,11 @@
 from taichi.core import ti_core as _ti_core
-from taichi.core.primitive_types import f32, u8
 from taichi.lang.impl import default_cfg, field, static
 from taichi.lang.kernel_arguments import ext_arr, template
 from taichi.lang.kernel_impl import kernel
 from taichi.lang.matrix import Vector
 from taichi.lang.ndrange import ndrange
 from taichi.lang.ops import atomic_add, get_addr
+from taichi.types.primitive_types import f32, u8
 
 import taichi as ti
